Changed types for db*ids from unsigned long to unsigned long long to be compatible with 64bit sqlite3 IDs.

+++ dbexchange.h
@@ -17,33 +17,34 @@ int setconfigsetting(char *setting, char *value);
 char *getsettingdata(char *setting);
 int writecsview(char *setting);
 int preparechannelstatement(); /* Run this before adding any channels */
-unsigned long addchannel(chanpropnode *achan, char *chanurl, 
-                         char *rss_version, char *directory);
+unsigned long long addchannel(chanpropnode *achan, char *chanurl, 
+                              char *rss_version, char *directory);
 int finalisechannelstatement(); /* Run this after adding all channels */
 int preparechancatstatement(); /* Run this before adding any channel categories */
-int addchancat(unsigned long channelid, catnode *acat);
+int addchancat(unsigned long long channelid, catnode *acat);
 void finalisechancatstatement(); /* Run this after adding all channel categories */
 int checkupdatechannel(chanpropnode *achan, char *rss_version);
 int prepareitemcatstatement(); /* Run this before adding any item categories */
-int additemcat(unsigned long itemid, catnode *acat);
+int additemcat(unsigned long long itemid, catnode *acat);
 void finaliseitemcatstatement(); /* Run this after adding all item categories */
 int prepareitemstatementdownloaded(); /* Run this before adding any dl items */
 int prepareitemstatementnotdownloaded(); /* Run this before adding any non-dl items */
-unsigned long additemdled(itempropnode *anitem, unsigned long chanid, 
-                          char *ofn, char *thefn); /* Add a Downloaded item */
-unsigned long additemndled(itempropnode *anitem, unsigned long chanid, char *ofn); /* Add a Non-Downloaded item */
+unsigned long long additemdled(itempropnode *anitem, unsigned long long chanid, 
+                               char *ofn, char *thefn); /* Add a Downloaded item */
+unsigned long long additemndled(itempropnode *anitem, unsigned long long chanid, 
+                                char *ofn); /* Add a Non-Downloaded item */
 void finaliseitemdlstatement(); /* Run this after adding all the downloaded items */
 void finaliseitemndlstatement(); /* Run this after adding all the non-downloaded items */
 int prepareuditemdlstatement(); /* Run this before updating any item to be downloaded */
-int updateitemdownloaded(char *filename, unsigned long itemid); /* Update item when it's been downloaded */
+int updateitemdownloaded(char *filename, unsigned long long itemid); /* Update item when it's been downloaded */
 void finaliseuditemdlstatement(); /* Run this after updating all items with download information */
 int prepareselitemstatement(); /* Run this before checking if the item exists */
-int isitemnew(itempropnode *anitem, unsigned long dbchannelid); /* Returns -1 on error, otherwise 1=true and 0=false */
+int isitemnew(itempropnode *anitem, unsigned long long dbchannelid); /* Returns -1 on error, otherwise 1=true and 0=false */
 void finaliseselitemstatement(); /* Run this after checking all items for newness */
-rssenclosure *getlatestitemenc(unsigned long chanid); /* Returns NULL on OoM or DB Error, or empty enc if none in DB */
-char *getlatestitemofn(unsigned long chanid); /* Returns NULL on OoM, blank str on no OFN */
-unsigned long getlatestitemid(unsigned long chanid);
-unsigned long getchanidfromurl(char *chanurl);
+rssenclosure *getlatestitemenc(unsigned long long chanid); /* Returns NULL on OoM or DB Error, or empty enc if none in DB */
+char *getlatestitemofn(unsigned long long chanid); /* Returns NULL on OoM, blank str on no OFN */
+unsigned long long getlatestitemid(unsigned long long chanid);
+unsigned long long getchanidfromurl(char *chanurl);
 static int callback_glio(void *NotUsed, int argc, char **argv, char **azColName);
 static int callback_glie(void *anenc, int argc, char **argv, char **azColName);
 static int callback_cuc(void *achan, int argc, char **argv, char **azColName);
